[SCRAPING] Package improvements; Separation of classes; Readme
