+++ .gitignore
@@ -1,0 +1,4 @@
+# created by virtualenv automatically
+./site-packages
+./results
+*.cfg
