+++ README.md
@@ -1,1 +1,13 @@
-# web-scraping-car
+# Web-Scraping-Car
+It´s one python web-scraping using beautifulSoup. The main objective
+is find latest cars ads in the site. Initially there is an implementation
+to find in the Icarros.
+
+> - The file `src/main.py` contains one example of use;
+> - The scraping result is store on folder `results` at root of project;
+> - This project uses `python 3.11`.
+
+## Commands to install required dependencies
+
+- `pip install requests` 
+- `pip install beautifulSoup4`
