+++ src/scraping/scraping.py
@@ -1,0 +1,8 @@
+import requests
+
+
+class Scraping:
+
+    def search(self, site_url):
+        scraping_content = requests.get(site_url).content
+        return scraping_content
